% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_bayesian.R
\docType{methods}
\name{CreateAndRunModel,StandardBayesian-method}
\alias{CreateAndRunModel,StandardBayesian-method}
\title{Testaa, onko jo tallennettuna t채m채n mallin pohjana olevasta datasta rds-tiedostoa. Jos ei ole, tallentaa sellaisen.}
\usage{
\S4method{CreateAndRunModel}{StandardBayesian}(object = StandardBayesian(),
  jags_settings = list(), ...)
}
\arguments{
\item{jags_settings}{lista jagsille annettavista asetuksista (adapt,n.chains,burnin,thin,method)}
}
\description{
Testaa, onko jo tallennettuna t채m채n mallin pohjana olevasta datasta rds-tiedostoa. Jos ei ole, tallentaa sellaisen.
}
