% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_bayesian_results.R
\name{GetAllPlots}
\alias{GetAllPlots}
\title{Luo caterpillarkuviot kaikista sijainneista sek채 lis채ksi interaktiosta 
kielen kanssa}
\usage{
GetAllPlots(varname.in.model, varname.in.df, vaihe1.df, post, sumstats)
}
\description{
Luo caterpillarkuviot kaikista sijainneista sek채 lis채ksi interaktiosta 
kielen kanssa
}
