% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{GetAsvo}
\alias{GetAsvo}
\title{A 'depsearcheR' filter for getting certain kinds of ASVO sentences}
\usage{
GetAsvo(sent)
}
\arguments{
\item{sent}{input sentence in conll format}
}
\description{
A 'depsearcheR' filter for getting certain kinds of ASVO sentences
}
