% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{headverbs}
\alias{headverbs}
\title{Frequency table of head verbs having count > 5}
\format{A tibble with n rows and y variables:
\describe{
  \item{lang}{language of the verb}
  \item{headverb}{lemma of the verb}
  \item{n}{frequency of the lemma}
  \item{is_presentational}{y(es) / n(o)}
  ...
}}
\usage{
headverbs
}
\description{
A dataset for determining wether or not a sentence with an adverbial has a 'presentational' main verb (eg. publish)
}
\keyword{datasets}
