% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_bayesian.R
\docType{class}
\name{StandardBayesian-class}
\alias{StandardBayesian-class}
\title{Tutkimuksessa standardisti käytetty bayesiläinen malli, jossa selitettävänä muuttujana sijainti,
ja lisäksi tutkitaan kielen ja muiden prediktoreiden välisiä interaktioita.}
\arguments{
\item{original.data}{alkuperäinen dataframe, joka on mallin aineistona}

\item{filename}{nimi tai nimen alku rds-tiedostoille, joihin tallennetaan aineisto ja posteriorijakauma}

\item{dump.path}{kansio, jonne rds-tiedostot tallennetaan}

\item{model.path}{kansio, jonne luodaan bugs-malli ja siihen liittyvä käynnistysskripti}

\item{source.file}{polku alkuperäisen aineiston säilyttävään rds-tiedostoon}

\item{post.file}{polku posteriorijakauman säilyttävään rds-tiedostoon}

\item{predictorlist}{lista, jossa on tiedot myös siitä, kuinka moniarvoisia muuttujia prediktorit ovat}

\item{datalist}{lista, johon on tallennettuna posteriorijakauma sekä useita kuvioita yms.}
}
\description{
Tutkimuksessa standardisti käytetty bayesiläinen malli, jossa selitettävänä muuttujana sijainti,
ja lisäksi tutkitaan kielen ja muiden prediktoreiden välisiä interaktioita.
}
