% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_bayesian.R
\name{StandardBayesian}
\alias{StandardBayesian}
\title{Konstruktori StandardBayesian-luokalle.}
\usage{
StandardBayesian(original.data = data.frame(), filename = character(),
  predictors = vector(), ...)
}
\description{
Konstruktori StandardBayesian-luokalle.
}
