% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reading.R
\name{TibbleFromJson}
\alias{TibbleFromJson}
\title{Reads a folder of json files to a tibble}
\usage{
TibbleFromJson(sourcefolder)
}
\arguments{
\item{sourcefolder}{the folder containing the files (must end with '/')}
}
\description{
Reads a folder of json files to a tibble
}
